no more explicit colormodes
