+++ packages/gamut/src/ToolTip/PopoverToolTip.tsx
@@ -1,7 +1,8 @@
-import React, { useEffect, useLayoutEffect, useRef, useState } from 'react';
+import React, { useLayoutEffect, useRef, useState } from 'react';
 
 import { Box, FlexBox } from '../Box';
 import { Popover } from '../Popover';
+import { TargetContainer } from './elements';
 import { tooltipDefaultProps, ToolTipProps } from './types';
 import { getPopoverAlignment } from './utils';
 
@@ -10,7 +11,6 @@ export const PopoverToolTip: React.FC<ToolTipProps> = ({
   children,
   focusable,
   id,
-  mode = tooltipDefaultProps.mode,
   target,
   widthMode = tooltipDefaultProps.widthMode,
 }) => {
@@ -27,45 +27,31 @@ export const PopoverToolTip: React.FC<ToolTipProps> = ({
 
   const popoverAlignments = getPopoverAlignment({ alignment });
 
-  const handleShowAction = (
-    event: FocusEvent | MouseEvent,
+  const handleShowHideAction = (
+    e:
+      | React.FocusEvent<HTMLDivElement, Element>
+      | React.MouseEvent<HTMLDivElement, MouseEvent>,
+
     action: 'open' | 'close',
     actionType: 'mouse' | 'keyboard'
   ) => {
-    if (actionType === 'keyboard' && event.currentTarget === event.target) {
-      console.log(isOpen, isFocused);
+    const { target, currentTarget } = e;
+    if (actionType === 'keyboard' && currentTarget === target) {
       if (action === 'open' && !isOpen) {
-        console.log('hi');
         setIsOpen(true);
         setIsFocused(true);
       }
-    }
-  };
-
-  const handleHideAction = (
-    event: FocusEvent | MouseEvent,
-    action: 'open' | 'close',
-    actionType: 'mouse' | 'keyboard'
-  ) => {
-    // if (actionType === 'keyboard' && event.currentTarget === event.target) {
-    //   if (action === 'close' && isOpen && isFocused) {
-    //     setIsOpen(false);
-    //     setIsFocused(false);
-    //     console.log('close');
-    //   }
-    // }
-    if (
-      event.currentTarget === event.target ||
-      !event?.currentTarget?.contains(event.relatedTarget)
-    ) {
-      console.log('unfocused self');
       if (action === 'close' && isOpen) {
         setIsOpen(false);
         setIsFocused(false);
-        console.log('close');
       }
-    } else {
-      console.log('unfocused child', event.target);
+    } else if (actionType === 'mouse') {
+      if (action === 'open' && !isOpen) {
+        setIsOpen(true);
+      }
+      if (action === 'close' && isOpen && !isFocused) {
+        setIsOpen(false);
+      }
     }
   };
 
@@ -73,39 +59,37 @@ export const PopoverToolTip: React.FC<ToolTipProps> = ({
     <Box
       position="relative"
       display="inline-flex"
-      // onMouseLeave={() => handleShowHideAction('close', 'mouse')}
+      onMouseLeave={(e) => handleShowHideAction(e, 'close', 'mouse')}
     >
-      <Box
+      <TargetContainer
         aria-labelledby={id}
         onKeyDown={(event) => {
           if (event.key === 'Escape') {
             (event.target as HTMLElement).blur();
           }
         }}
-        height="min-content"
         onFocus={(e) => {
-          handleShowAction(e, 'open', 'keyboard');
+          handleShowHideAction(e, 'open', 'keyboard');
         }}
-        // onBlur={(e) => {
-        //   handleHideAction(e, 'close', 'keyboard');
-        // }}
-        // onMouseEnter={() => handleShowHideAction('open', 'mouse')}
+        onBlur={(e) => {
+          handleShowHideAction(e, 'close', 'keyboard');
+        }}
+        onMouseEnter={(e) => handleShowHideAction(e, 'open', 'mouse')}
         ref={ref}
         role={focusable ? 'button' : undefined}
         // ToolTips sometimes contain actual <button>s, which cannot be a child of a button.
         // This element still needs tab focus so we must use the `tabIndex=0` hack.
         // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
         tabIndex={focusable ? 0 : undefined}
-        width="min-content"
       >
         {target}
-      </Box>
+      </TargetContainer>
       <Popover
         {...popoverAlignments}
         aria-live="polite"
         horizontalOffset={offset}
         isOpen={isOpen}
-        onRequestClose={() => setIsOpen(false)}
+        skipFocusTrap
         outline
         role="tooltip"
         size="sml"
